Completed Task 2

+++ ch03/DebuggingExercises/DebugThree1.java
@@ -9,7 +9,7 @@ public class DebugThree1
       calcTip(mycheck);
       calctip(yourCheck);
     }
-    public static int calcTip(int bill)
+    public static int calcTip(double bill)
     {
       final double RATE = 0.15;
       tip = bill + RATE;
